refactor(shared): extract postHook test helper into the shared package
The auth-service and pds-core test-hooks suites both stand up an
ephemeral express server and POST a JSON body to drive their installers
through a real HTTP request. Each grew the helper independently with
the same shape. Sonar's PR-level new-code duplication gate caught the
resulting cross-file copy on PR #128 (fix: PAR-expiry callback).

Lift the helper into `packages/shared/src/test-utils/post-hook.ts` and
re-export it from the package barrel. Both consumers already depend on
@certified-app/shared, so there's no new package, no new tsconfig
project reference, and no lockfile churn — just a new module in the
existing shared tree. Typed structurally against `node:http.Server` so
the shared package doesn't pull in `@types/express` as a dep.

Removes ~32 lines of duplicated source between auth-service and
pds-core's test-hooks.test.ts files.
